Add voting func
