+++ client/public/index.html
@@ -3,6 +3,7 @@
   <head>
     <meta charset="UTF-8" />
     <meta name="viewport" content="width=device-width, initial-scale=1, shrink-to-fit=no" />
+    <script src="https://kit.fontawesome.com/b77f0cfc19.js" crossorigin="anonymous"></script>
     <title>Engage Starter</title>
   </head>
